requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0

lxml~=4.9.3
bs4~=0.0.1
soupsieve~=2.5
pip~=23.0
Pillow~=9.5.0
tornado~=6.2
cssselect~=1.2.0
pyee~=8.2.2
colorama~=0.4.6
w3lib~=2.1.2
click~=8.1.3
certifi~=2022.12.7
pyquery~=2.0.0
idna~=3.4
urllib3~=1.26.14
simplejson~=3.19.2
pyppeteer~=1.0.2
appdirs~=1.4.4
websockets~=10.4
tqdm~=4.66.1
setuptools~=65.5.0
packaging~=23.1
pyparsing~=3.0.9
zipp~=3.17.0
parse~=1.19.1
python-dotenv~=0.21.1